upgrade the examples to Kamon 1.0
