+++ .gitignore
@@ -4,6 +4,7 @@
 *.sc
 .pygments-cache
 .DS_Store
+.g8
 
 # sbt specific
 dist/*
