+++ examples/akka-ping-pong-with-kamon/project/plugins.sbt
@@ -1,0 +1,2 @@
+resolvers += Resolver.bintrayRepo("kamon-io", "sbt-plugins")
+addSbtPlugin("io.kamon" % "sbt-aspectj-runner" % "1.1-73454753f2290789913ff1f4bba8a620c7142f77")
