+++ examples/akka-ping-pong-with-kamon/src/main/resources/application.conf
@@ -1,0 +1,11 @@
+kamon {
+  environment {
+    service = "akka-ping-pong"
+  }
+
+  util.filters {
+    "akka.tracked-actor" {
+      includes = [ "**" ]
+    }
+  }
+}
