--- examples/kamon-play/app/Filters.scala
@@ -1,24 +1,0 @@
-/* ===================================================
- * Copyright © 2013-2015 the kamon project <http://kamon.io/>
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * ========================================================== */
-
-
-import javax.inject.Inject
-import play.api.http.HttpFilters
-import filters.TraceLocalFilter
-
-class Filters @Inject() (traceLocalFilter: TraceLocalFilter)  extends HttpFilters {
-  val filters = Seq(traceLocalFilter)
-}
