--- examples/kamon-play/conf/application.conf
@@ -1,42 +1,0 @@
-#kamon related configuration
-
-kamon {
-  metric {
-    tick-interval = 1 second
-  }
-
-  play {
-    include-trace-token-header = true
-    trace-token-header-name = "X-Trace-Token"
-  }
-
-  metric {
-    filters {
-      trace.includes = [ "**" ]
-      akka-actor.includes = [ "**" ]
-      akka-actor.excludes = [ ]
-      akka-dispatcher.excludes = [ ]
-      akka-dispatcher.includes = [ "**" ]
-    }
-  }
-
-  modules {
-    kamon-log-reporter.auto-start = no
-  }
-}
-
-# This is the main configuration file for the application.
-# ~~~~~
-
-# Secret key
-# ~~~~~
-# The secret key is used to secure cryptographics functions.
-# If you deploy your application to several instances be sure to use the same key!
-application.secret = "3BLM`<aD^5r/L[MinNdw8Tp@915n0djY[g66OSOLi@?k`>AZE9EOphrmf;;6JsAN"
-
-# The application languages
-# ~~~~~
-application.langs = "en"
-
-# HttpRequestHandler
-play.http.requestHandler = "play.http.DefaultHttpRequestHandler"
