--- examples/kamon-play/conf/routes
@@ -1,5 +1,0 @@
-# Routes
-GET     /helloKamon                      @controllers.KamonPlayExample.sayHello
-GET     /helloKamonWithTraceName         @controllers.KamonPlayExample.sayHelloWithTraceName
-GET     /incrementCounter                @controllers.KamonPlayExample.incrementCounter
-GET     /updateTraceLocal                @controllers.KamonPlayExample.updateTraceLocal
