--- examples/ping-pong/src/main/resources/application.conf
@@ -1,18 +1,0 @@
-kamon {
-
-  trace {
-    level = simple-trace
-  }
-
-  metric {
-    filters {
-      trace.includes = [ "**" ]
-      akka-actor.includes = [ "**" ]
-      akka-dispatcher.includes = [ "**" ]
-    }
-  }
-
-  modules {
-    kamon-log-reporter.auto-start = yes
-  }
-}
