--- examples/ping-pong/src/main/resources/logback.xml
@@ -1,11 +1,0 @@
-<configuration scan="true">
-    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder>
-            <pattern>%date{HH:mm:ss.SSS} %-5level [%X{akkaSource}] [%thread] %logger{55} - %msg%n</pattern>
-        </encoder>
-    </appender>
-
-    <root level="debug">
-        <appender-ref ref="STDOUT" />
-    </root>
-</configuration>
