--- examples/ping-pong/src/test/scala/kamon/aspectj/sbt/runner/PingPongTest.scala
@@ -1,37 +1,0 @@
-/*
- * =========================================================================================
- * Copyright © 2013-2015 the kamon project <http://kamon.io/>
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
- * except in compliance with the License. You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the
- * License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific language governing permissions
- * and limitations under the License.
- * =========================================================================================
- */
-
-package com.typesafe.echo.sample
-
-import akka.actor._
-import kamon.Kamon
-
-object PingPongTest extends App {
-  Kamon.start()
-
-  val system = ActorSystem()
-
-  val pinger = system.actorOf(Props(new Actor {
-    def receive: Actor.Receive = { case "pong" ⇒ sender ! "ping" }
-  }), "pinger")
-
-  val ponger = system.actorOf(Props(new Actor {
-    def receive: Actor.Receive = { case "ping" ⇒ sender ! "pong" }
-  }), "ponger")
-
-  pinger.tell("pong", ponger)
-}
-
