+++ examples/play-scala-seed-with-kamon/.gitignore
@@ -1,0 +1,8 @@
+logs
+target
+/.idea
+/.idea_modules
+/.classpath
+/.project
+/.settings
+/RUNNING_PID
