+++ examples/play-scala-seed-with-kamon/app/controllers/HomeController.scala
@@ -1,0 +1,27 @@
+package controllers
+
+import javax.inject._
+import play.api._
+import play.api.mvc._
+import kamon.Kamon
+
+/**
+ * This controller creates an `Action` to handle HTTP requests to the
+ * application's home page.
+ */
+@Singleton
+class HomeController @Inject()(cc: ControllerComponents) extends AbstractController(cc) {
+
+  /**
+   * Create an Action to render an HTML page.
+   *
+   * The configuration in the `routes` file means that this method
+   * will be called when the application receives a `GET` request with
+   * a path of `/`.
+   */
+  def index() = Action { implicit request: Request[AnyContent] =>
+    // If you see the TraceID in the console then instrumentation must be working.
+    println("Current Trace ID: " + Kamon.currentSpan().context().traceID.string)
+    Ok(views.html.index())
+  }
+}
