+++ examples/play-scala-seed-with-kamon/app/views/index.scala.html
@@ -1,0 +1,5 @@
+@()
+
+@main("Welcome to Play") {
+  <h1>Welcome to Play!</h1>
+}
