+++ examples/play-scala-seed-with-kamon/build.sbt
@@ -1,0 +1,18 @@
+name := """play-scala-seed-with-kamon"""
+organization := "io.kamon"
+
+version := "1.0-SNAPSHOT"
+
+lazy val root = (project in file(".")).enablePlugins(PlayScala)
+
+scalaVersion := "2.12.3"
+
+libraryDependencies += guice
+libraryDependencies += "io.kamon" %% "kamon-play-2.6" % "1.0.1"
+libraryDependencies += "org.scalatestplus.play" %% "scalatestplus-play" % "3.1.2" % Test
+
+// Adds additional packages into Twirl
+//TwirlKeys.templateImports += "io.kamon.controllers._"
+
+// Adds additional packages into conf/routes
+// play.sbt.routes.RoutesKeys.routesImport += "io.kamon.binders._"
