+++ examples/play-scala-seed-with-kamon/conf/application.conf
@@ -1,0 +1,1 @@
+# https://www.playframework.com/documentation/latest/Configuration
