+++ examples/play-scala-seed-with-kamon/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,5 @@
+distributionUrl=https\://services.gradle.org/distributions/gradle-4.4-bin.zip
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStorePath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
