+++ examples/play-scala-seed-with-kamon/project/scaffold.sbt
@@ -1,0 +1,5 @@
+// Defines scaffolding (found under .g8 folder)
+// http://www.foundweekends.org/giter8/scaffolding.html
+// sbt "g8Scaffold form"
+// not working on sbt 1.0
+//addSbtPlugin("org.foundweekends.giter8" % "sbt-giter8-scaffold" % "0.10.0")
