+++ examples/play-scala-seed-with-kamon/public/images/favicon.png
